Removes Action suffix to respect Symfony Controller Best Practices
