+++ src/UI/Action/Phone/CreatePhone.php
@@ -1,36 +1,34 @@
 <?php
 declare(strict_types=1);
 
-namespace App\UI\Action\User;
+namespace App\UI\Action\Phone;
 
 use App\App\Validator\ApiValidator;
 use App\App\ErrorException\ApiError;
 use App\App\ErrorException\ApiException;
 use App\App\Validator\Interfaces\ApiValidatorInterface;
-use App\Domain\DTO\UserDTO;
-use App\Domain\Entity\User;
-use App\Domain\Repository\UserRepository;
-use App\UI\Action\User\Interfaces\CreateUserActionInterface;
+use App\Domain\Entity\Phone;
+use App\Domain\Repository\PhoneRepository;
+use App\UI\Action\Phone\Interfaces\CreatePhoneInterface;
 use App\UI\Responder\Interfaces\CreateResponderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Symfony\Component\Serializer\Exception\NotEncodableValueException;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @Route(
- *     "/api/users",
- *     name="user_create",
+ *     "/api/phones",
+ *     name="phone_create",
  *     methods={"POST"}
  * )
  *
- * Class CreateUserAction
- * @package App\UI\Action\User
+ * Class CreatePhoneAction
+ * @package App\UI\Action\Phone
  */
-final class CreateUserAction implements CreateUserActionInterface
+final class CreatePhone implements CreatePhoneInterface
 {
     /**
      * @var SerializerInterface
@@ -43,9 +41,9 @@ final class CreateUserAction implements CreateUserActionInterface
     private $apiValidator;
 
     /**
-     * @var UserRepository
+     * @var PhoneRepository
      */
-    private $userRepository;
+    private $phoneRepository;
 
     /**
      * @var UrlGeneratorInterface
@@ -53,25 +51,18 @@ final class CreateUserAction implements CreateUserActionInterface
     private $urlGenerator;
 
     /**
-     * @var UserPasswordEncoderInterface
-     */
-    private $passwordEncoder;
-
-    /**
      * {@inheritdoc}
      */
     public function __construct(
         SerializerInterface $serializer,
         ApiValidatorInterface $apiValidator,
-        UserRepository $userRepository,
-        UrlGeneratorInterface $urlGenerator,
-        UserPasswordEncoderInterface $passwordEncoder
+        PhoneRepository $phoneRepository,
+        UrlGeneratorInterface $urlGenerator
     ) {
         $this->serializer = $serializer;
         $this->apiValidator = $apiValidator;
-        $this->userRepository = $userRepository;
+        $this->phoneRepository = $phoneRepository;
         $this->urlGenerator = $urlGenerator;
-        $this->passwordEncoder = $passwordEncoder;
     }
 
 
@@ -83,30 +74,20 @@ final class CreateUserAction implements CreateUserActionInterface
         $json = $request->getContent();
 
         try {
-            $userDTO = $this->serializer->deserialize($json, UserDTO::class, 'json');
+            $phone = $this->serializer->deserialize($json, Phone::class, 'json');
         } catch (NotEncodableValueException $e) {
             $apiError = new ApiError(Response::HTTP_BAD_REQUEST, ApiError::TYPE_INVALID_REQUEST_BODY_FORMAT);
             throw new ApiException($apiError);
         }
 
-        $user = new User();
-        $password = $this->passwordEncoder->encodePassword($user, $userDTO->password);
+        $this->apiValidator->validate($phone, null, ['phone']);
 
-        $user->registration(
-            $userDTO->username,
-            $password,
-            $userDTO->email,
-            $userDTO->phoneNumber
-        );
+        $this->phoneRepository->save($phone);
 
-        $this->apiValidator->validate($user, null, ['user']);
-
-        $this->userRepository->save($user);
-
-        $jsonUser = $this->serializer->serialize($user, 'json', [
-            'groups' => ['user']
+        $jsonPhone = $this->serializer->serialize($phone, 'json', [
+            'groups' => ['phone']
         ]);
 
-        return $responder($jsonUser, $user);
+        return $responder($jsonPhone, $phone);
     }
 }
