+++ src/UI/Action/Phone/UpdatePhone.php
@@ -3,16 +3,17 @@ declare(strict_types=1);
 
 namespace App\UI\Action\Phone;
 
-use App\App\Validator\ApiValidator;
 use App\App\ErrorException\ApiError;
 use App\App\ErrorException\ApiException;
 use App\App\Validator\Interfaces\ApiValidatorInterface;
-use App\Domain\Entity\Phone;
+use App\Domain\DTO\PhoneDTO;
 use App\Domain\Repository\PhoneRepository;
-use App\UI\Action\Phone\Interfaces\CreatePhoneActionInterface;
-use App\UI\Responder\Interfaces\CreateResponderInterface;
+use App\UI\Action\Phone\Interfaces\UpdatePhoneInterface;
+use App\UI\Responder\Interfaces\UpdateResponderInterface;
+use App\UI\Responder\UpdateResponder;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Serializer\Exception\NotEncodableValueException;
@@ -20,15 +21,15 @@ use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @Route(
- *     "/api/phones",
- *     name="phone_create",
- *     methods={"POST"}
+ *     "/api/phones/{id}",
+ *     name="phone_update",
+ *     methods={"PUT"}
  * )
  *
- * Class CreatePhoneAction
+ * Class UpdatePhoneAction
  * @package App\UI\Action\Phone
  */
-final class CreatePhoneAction implements CreatePhoneActionInterface
+final class UpdatePhone implements UpdatePhoneInterface
 {
     /**
      * @var SerializerInterface
@@ -36,7 +37,7 @@ final class CreatePhoneAction implements CreatePhoneActionInterface
     private $serializer;
 
     /**
-     * @var ApiValidator
+     * @var ApiValidatorInterface
      */
     private $apiValidator;
 
@@ -65,28 +66,39 @@ final class CreatePhoneAction implements CreatePhoneActionInterface
         $this->urlGenerator = $urlGenerator;
     }
 
-
     /**
      * {@inheritdoc}
      */
-    public function __invoke(Request $request, CreateResponderInterface $responder): Response
+    public function __invoke(Request $request, UpdateResponderInterface $responder): Response
     {
         $json = $request->getContent();
 
+        $phoneId = $request->attributes->get('id');
+
+        $phone = $this->phoneRepository->findOneById($phoneId);
+
+        if (!$phone) {
+            throw new NotFoundHttpException(sprintf('Resource %s not found with id "%s"', 'Phone', $phoneId));
+        }
+
         try {
-            $phone = $this->serializer->deserialize($json, Phone::class, 'json');
+            $phoneDTO = $this->serializer->deserialize($json, PhoneDTO::class, 'json');
         } catch (NotEncodableValueException $e) {
             $apiError = new ApiError(Response::HTTP_BAD_REQUEST, ApiError::TYPE_INVALID_REQUEST_BODY_FORMAT);
             throw new ApiException($apiError);
         }
 
+        $phone->update(
+            $phoneDTO->description,
+            $phoneDTO->price,
+            $phoneDTO->stock
+        );
+
         $this->apiValidator->validate($phone, null, ['phone']);
 
         $this->phoneRepository->save($phone);
 
-        $jsonPhone = $this->serializer->serialize($phone, 'json', [
-            'groups' => ['phone']
-        ]);
+        $jsonPhone = $this->serializer->serialize($phone, 'json', ['groups' => ['phone']]);
 
         return $responder($jsonPhone, $phone);
     }
