+++ src/UI/Action/User/ReadUser.php
@@ -4,25 +4,32 @@ declare(strict_types=1);
 namespace App\UI\Action\User;
 
 use App\Domain\Repository\UserRepository;
-use App\UI\Action\User\Interfaces\ReadUserListActionInterface;
+use App\UI\Action\User\Interfaces\ReadUserInterface;
 use App\UI\Responder\Interfaces\ReadResponderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @Route(
- *     "/api/users",
- *     name="users_list",
+ *     "/api/users/{id}",
+ *     name="user_read",
  *     methods={"GET"}
  * )
  *
- * Class UserAction
+ * Class ReadUserAction
  * @package App\UI\Action\User
  */
-final class ReadUserListAction implements ReadUserListActionInterface
+final class ReadUser implements ReadUserInterface
 {
+    /**
+     * @var AuthorizationCheckerInterface
+     */
+    private $authorizationChecker;
+
     /**
      * @var UserRepository
      */
@@ -37,9 +44,11 @@ final class ReadUserListAction implements ReadUserListActionInterface
      * {@inheritdoc}
      */
     public function __construct(
+        AuthorizationCheckerInterface $authorizationChecker,
         UserRepository $userRepository,
         SerializerInterface $serializer
     ) {
+        $this->authorizationChecker = $authorizationChecker;
         $this->userRepository = $userRepository;
         $this->serializer = $serializer;
     }
@@ -49,9 +58,16 @@ final class ReadUserListAction implements ReadUserListActionInterface
      */
     public function __invoke(Request $request, ReadResponderInterface $responder): Response
     {
-        $timestamp = $this->userRepository->getLastUpdateDate();
+        $userId = $request->attributes->get('id');
+
+        $userLastUpdateDate = $this->userRepository->getOneUpdateDate($userId);
+
+        if (!$userLastUpdateDate) {
+            throw new NotFoundHttpException(sprintf('Resource %s not found with id "%s"', 'User', $userId));
+        }
+
         $lastModified = new \DateTime();
-        $lastModified->setTimestamp(intval($timestamp));
+        $lastModified->setTimestamp(intval($userLastUpdateDate));
 
         $response = new Response();
         $response->setLastModified($lastModified);
@@ -61,11 +77,11 @@ final class ReadUserListAction implements ReadUserListActionInterface
             return $response;
         }
 
-        $users = $this->userRepository->findAll();
+        $user = $this->userRepository->findOneById($userId);
 
-        $json  = $this->serializer->serialize($users, 'json', ['groups' => ['users_list']]);
+        $json = $this->serializer->serialize($user, 'json', ['groups' => ['user']]);
 
-        $response->setStatusCode(200);
+        $response->setStatusCode(Response::HTTP_OK);
         return $response->setContent($json);
 
 //        return $responder($json);
